elasticrepoused

+++ src/main/java/com/massil/dto/FilterDropdowns.java
@@ -16,7 +16,7 @@ public class FilterDropdowns extends Response {
     private List<String> make;
     private List<String> model;
     private List<String> series;
-    private List<Integer> year;
+    private List<Long> year;
     private List<String> engine;
     private List<String> transmission;
